speech_quality_prediction/requirements.txt
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1
librosa==0.10.1
scikit-learn==1.2.2
matplotlib==3.7.1
PyYAML==6.0
torch==2.0.1
torchaudio==2.0.2
pesq==0.0.4
pystoi==0.3.3
tensorboard==2.13.0
soundfile==0.12.1 